+++ packages/EDA/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.5.6 (2026-06-01)
 
-Add two-sample t-test (Welch and Student)
+* Add two-sample t-test (Welch and Student)
+* Add Control comparisons (Dunnett's test and Holm-corrected Welch's t-tests)
 
 ## 1.5.5 (2026-05-26)